+++ tests/PhpDATest/Entity/AnalysisCollectionTest.php
@@ -46,6 +46,13 @@ class AnalysisCollectionTest extends \PHPUnit_Framework_TestCase
         $this->assertSame($this->graph, $this->fixture->getGraph());
     }
 
+    public function testMutateAndAccessCallMode()
+    {
+        $this->assertFalse($this->fixture->isCallMode());
+        $this->fixture->setCallMode();
+        $this->assertTrue($this->fixture->isCallMode());
+    }
+
     public function testMutateAndAccessLayout()
     {
         $this->assertInstanceOf('PhpDA\Layout\Null', $this->fixture->getLayout());
@@ -184,6 +191,111 @@ class AnalysisCollectionTest extends \PHPUnit_Framework_TestCase
         $declaredNameVertex->shouldReceive('hasEdgeTo')->with($usedName1Vertex)->andReturn(false);
         $declaredNameVertex->shouldReceive('createEdgeTo')->with($usedName1Vertex)->once()->andReturn($edge);
         $declaredNameVertex->shouldReceive('hasEdgeTo')->with($usedName2Vertex)->andReturn(true);
+
+        $declaredNameVertex->shouldReceive('hasEdgeTo')->with($unsupportedName1Vertex)->andReturn(false);
+        $declaredNameVertex->shouldReceive('createEdgeTo')->with($unsupportedName1Vertex)->once()->andReturn(
+            $edgeUnsupported
+        );
+        $declaredNameVertex->shouldReceive('hasEdgeTo')->with($unsupportedName2Vertex)->andReturn(true);
+
+        $declaredNameVertex->shouldReceive('hasEdgeTo')->with($stringName1Vertex)->andReturn(false);
+        $declaredNameVertex->shouldReceive('createEdgeTo')->with($stringName1Vertex)->once()->andReturn(
+            $edgeNamespaced
+        );
+        $declaredNameVertex->shouldReceive('hasEdgeTo')->with($stringName2Vertex)->andReturn(false);
+        $declaredNameVertex->shouldReceive('createEdgeTo')->with($stringName2Vertex)->once()->andReturn(
+            $edgeNamespaced
+        );
+
+        $analysis = \Mockery::mock('PhpDA\Entity\Analysis');
+        $analysis->shouldReceive('getAdts')->andReturn(array($adt));
+        $analysis->shouldReceive('hasParseError')->once()->andReturn(false);
+        $this->fixture->attach($analysis);
+
+        $this->assertFalse($this->fixture->hasAnalysisFailures());
+        $this->assertEmpty($this->fixture->getAnalysisFailures());
+    }
+
+    public function testAttachAnalysisForCallMode()
+    {
+        $this->fixture->setCallMode();
+
+        $vertexLayout = array(1, 1);
+        $vertexUnsupportedLayout = array(1, 1);
+        $vertexNamespacedStringLayout = array(1, 1);
+        $edgeLayout = array(2, 1);
+        $edgeUnsupportedLayout = array(2, 5);
+        $edgeNamespacedStringLayout = array(2, 6);
+        $layout = \Mockery::mock('PhpDA\Layout\LayoutInterface');
+        $layout->shouldReceive('getVertex')->andReturn($vertexLayout);
+        $layout->shouldReceive('getVertexNamespacedString')->andReturn($vertexNamespacedStringLayout);
+        $layout->shouldReceive('getVertexUnsupported')->andReturn($vertexUnsupportedLayout);
+        $layout->shouldReceive('getEdge')->andReturn($edgeLayout);
+        $layout->shouldReceive('getEdgeUnsupported')->andReturn($edgeUnsupportedLayout);
+        $layout->shouldReceive('getEdgeNamespacedString')->andReturn($edgeNamespacedStringLayout);
+        $this->fixture->setLayout($layout);
+
+        $adt = \Mockery::mock('PhpDA\Entity\Adt');
+
+        $declaredName = \Mockery::mock('PhpParser\Node\Name');
+        $declaredName->shouldReceive('toString')->once()->andReturn('declaredName');
+        $adt->shouldReceive('getDeclaredNamespace')->once()->andReturn($declaredName);
+
+        $calledName1 = \Mockery::mock('PhpParser\Node\Name');
+        $calledName1->shouldReceive('toString')->once()->andReturn('calledName1');
+        $calledName2 = \Mockery::mock('PhpParser\Node\Name');
+        $calledName2->shouldReceive('toString')->once()->andReturn('calledName2');
+        $adt->shouldReceive('getCalledNamespaces')->once()->andReturn(array($calledName1, $calledName2));
+
+        $unsupportedName1 = \Mockery::mock('PhpParser\Node\Name');
+        $unsupportedName1->shouldReceive('toString')->once()->andReturn('unsupportedName1');
+        $unsupportedName2 = \Mockery::mock('PhpParser\Node\Name');
+        $unsupportedName2->shouldReceive('toString')->once()->andReturn('unsupportedName2');
+        $adt->shouldReceive('getUnsupportedStmts')->once()->andReturn(array($unsupportedName1, $unsupportedName2));
+
+        $stringName1 = \Mockery::mock('PhpParser\Node\Name');
+        $stringName1->shouldReceive('toString')->once()->andReturn('stringName1');
+        $stringName2 = \Mockery::mock('PhpParser\Node\Name');
+        $stringName2->shouldReceive('toString')->once()->andReturn('stringName2');
+        $adt->shouldReceive('getNamespacedStrings')->once()->andReturn(array($stringName1, $stringName2));
+
+        $declaredNameVertex = \Mockery::mock('Fhaculty\Graph\Vertex');
+        $declaredNameVertex->shouldReceive('setLayout')->with($vertexLayout)->once()->andReturnSelf();
+        $calledName1Vertex = \Mockery::mock('Fhaculty\Graph\Vertex');
+        $calledName1Vertex->shouldReceive('setLayout')->with($vertexLayout)->once()->andReturnSelf();
+        $calledName2Vertex = \Mockery::mock('Fhaculty\Graph\Vertex');
+        $calledName2Vertex->shouldReceive('setLayout')->with($vertexLayout)->once()->andReturnSelf();
+        $unsupportedName1Vertex = \Mockery::mock('Fhaculty\Graph\Vertex');
+        $unsupportedName1Vertex->shouldReceive('setLayout')->with($vertexLayout)->once()->andReturnSelf();
+        $unsupportedName1Vertex->shouldReceive('setLayout')->with($vertexUnsupportedLayout)->once()->andReturnSelf();
+        $unsupportedName2Vertex = \Mockery::mock('Fhaculty\Graph\Vertex');
+        $unsupportedName2Vertex->shouldReceive('setLayout')->with($vertexLayout)->once()->andReturnSelf();
+        $unsupportedName2Vertex->shouldReceive('setLayout')->with($vertexUnsupportedLayout)->once()->andReturnSelf();
+        $stringName1Vertex = \Mockery::mock('Fhaculty\Graph\Vertex');
+        $stringName1Vertex->shouldReceive('setLayout')->with($vertexLayout)->once()->andReturnSelf();
+        $stringName1Vertex->shouldReceive('setLayout')->with($vertexNamespacedStringLayout)->once()->andReturnSelf();
+        $stringName2Vertex = \Mockery::mock('Fhaculty\Graph\Vertex');
+        $stringName2Vertex->shouldReceive('setLayout')->with($vertexLayout)->once()->andReturnSelf();
+        $stringName2Vertex->shouldReceive('setLayout')->with($vertexNamespacedStringLayout)->once()->andReturnSelf();
+
+        $this->graph->shouldReceive('createVertex')->with('declaredName', true)->andReturn($declaredNameVertex);
+        $this->graph->shouldReceive('createVertex')->with('calledName1', true)->andReturn($calledName1Vertex);
+        $this->graph->shouldReceive('createVertex')->with('calledName2', true)->andReturn($calledName2Vertex);
+        $this->graph->shouldReceive('createVertex')->with('unsupportedName1', true)->andReturn($unsupportedName1Vertex);
+        $this->graph->shouldReceive('createVertex')->with('unsupportedName2', true)->andReturn($unsupportedName2Vertex);
+        $this->graph->shouldReceive('createVertex')->with('stringName1', true)->andReturn($stringName1Vertex);
+        $this->graph->shouldReceive('createVertex')->with('stringName2', true)->andReturn($stringName2Vertex);
+
+        $edge = \Mockery::mock('Fhaculty\Graph\Edge\Directed');
+        $edge->shouldReceive('setLayout')->with($edgeLayout);
+        $edgeUnsupported = \Mockery::mock('Fhaculty\Graph\Edge\Directed');
+        $edgeUnsupported->shouldReceive('setLayout')->with($edgeUnsupportedLayout);
+        $edgeNamespaced = \Mockery::mock('Fhaculty\Graph\Edge\Directed');
+        $edgeNamespaced->shouldReceive('setLayout')->with($edgeNamespacedStringLayout);
+
+        $declaredNameVertex->shouldReceive('hasEdgeTo')->with($calledName1Vertex)->andReturn(false);
+        $declaredNameVertex->shouldReceive('createEdgeTo')->with($calledName1Vertex)->once()->andReturn($edge);
+        $declaredNameVertex->shouldReceive('hasEdgeTo')->with($calledName2Vertex)->andReturn(true);
 
         $declaredNameVertex->shouldReceive('hasEdgeTo')->with($unsupportedName1Vertex)->andReturn(false);
         $declaredNameVertex->shouldReceive('createEdgeTo')->with($unsupportedName1Vertex)->once()->andReturn(