+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Change default configuration by adding visitorOptions for MetaNamespaceCollector
 - Add Meta object to Adt with corresponding collector
 - Change default configuration by adding mode config
-- Add mode configuration to switch analysis between call graph (default) and inheritance graph
+- Add mode configuration to switch analysis between usage graph (default), call graph and inheritance graph
+- Segregate call graph from analysis
 - Segregate inheritance graph from analysis
 - Add validation on opcode cache configuration for PhpDocBlocks TagCollector
 - Refactor Analyze Command and Service Factories